clean: more config cleanup and consolidation
