+++ client/src/www/app/outline_server_repository/vpn.ts
@@ -12,65 +12,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import * as net from '@outline/infrastructure/net';
-
-/**
- * getAddressFromTransportConfig returns the address of the tunnel server, if there's a meaningful one.
- * This is used to show the server address in the UI when connected.
- */
-export function getAddressFromTransportConfig(
-  transport: TransportConfigJson
-): string | undefined {
-  const hostConfig: {host?: string; port?: string} = transport;
-  if (hostConfig.host && hostConfig.port) {
-    return net.joinHostPort(hostConfig.host, hostConfig.port);
-  } else if (hostConfig.host) {
-    return hostConfig.host;
-  } else {
-    return undefined;
-  }
-}
-
-/**
- * getHostFromTransportConfig returns the host of the tunnel server, if there's a meaningful one.
- * This is used by the proxy resolution in Electron.
- */
-export function getHostFromTransportConfig(
-  transport: TransportConfigJson
-): string | undefined {
-  return (transport as unknown as {host: string | undefined}).host;
-}
-
-/**
- * setTransportConfigHost returns a new TransportConfigJson with the given host as the tunnel server.
- * Should only be set if getHostFromTransportConfig returns one.
- * This is used by the proxy resolution in Electron.
- */
-export function setTransportConfigHost(
-  transport: TransportConfigJson,
-  newHost: string
-): TransportConfigJson | undefined {
-  if (!('host' in transport)) {
-    return undefined;
-  }
-  return {...transport, host: newHost};
-}
+import {TunnelConfigJson} from './config';
 
 export const enum TunnelStatus {
   CONNECTED,
   DISCONNECTED,
   RECONNECTING,
   DISCONNECTING,
-}
-
-export type TransportConfigJson = object;
-
-/** TunnelConfigJson represents the configuration to set up a tunnel. */
-export interface TunnelConfigJson {
-  /** transport describes how to establish connections to the destinations.
-   * See https://github.com/Jigsaw-Code/outline-apps/blob/master/client/go/outline/config.go for format. */
-  transport: TransportConfigJson;
-  // This is the place where routing configuration would go.
 }
 
 /** StartRequestJson is the serializable request to start the VPN, used for persistence and IPCs. */
